Fix for headless systems
Without 'agg' parameter, matplotlib has an ultimate dependency on
XWindows XCB API. This library will not always be installed on server
systems. This fix allows the library to generate the plot png file and
inject it into the HTML IMG tag, without having to bring in some or all
of X as a dependency.